+++ core/unit_test/TestRangePolicyConstructors.hpp
@@ -17,6 +17,7 @@
 #include <gtest/gtest.h>
 
 #include <Kokkos_Core.hpp>
+
 namespace {
 
 TEST(TEST_CATEGORY, range_policy_runtime_parameters) {
@@ -68,118 +69,5 @@ TEST(TEST_CATEGORY, range_policy_runtime_parameters) {
     ASSERT_EQ(p1.chunk_size(), p2.chunk_size());
   }
 }
-
-struct TestRangePolicyCTADs {
-  struct SomeExecutionSpace {
-    using execution_space = SomeExecutionSpace;
-    using size_type       = size_t;
-  };
-  static_assert(Kokkos::is_execution_space_v<SomeExecutionSpace>);
-
-  struct ImplicitlyConvertibleToDefaultExecutionSpace {
-    operator Kokkos::DefaultExecutionSpace() const;
-  };
-  static_assert(!Kokkos::is_execution_space_v<
-                ImplicitlyConvertibleToDefaultExecutionSpace>);
-
-  static inline Kokkos::DefaultExecutionSpace des;
-  static inline ImplicitlyConvertibleToDefaultExecutionSpace notEs;
-  static inline SomeExecutionSpace ses;
-  static inline TEST_EXECSPACE es;
-
-  using RangePolicyTestExecSpace = std::conditional_t<
-      std::is_same_v<TEST_EXECSPACE, Kokkos::DefaultExecutionSpace>,
-      Kokkos::RangePolicy<>, Kokkos::RangePolicy<TEST_EXECSPACE> >;
-
-  static inline int64_t i64;
-  static inline int32_t i32;
-  static inline Kokkos::ChunkSize cs{0};
-
-  static inline Kokkos::RangePolicy<> rp;
-
-  //  Copy/move
-  //
-  // icpc generates an ICE on implicit deduction guides for copy/move
-  //
-  // internal error: assertion failed: find_placeholder_arg_for_pack: symbol
-  // not found (scope_stk.c, line 11248 in find_placeholder_arg_for_pack)
-
-#if not defined(__INTEL_COMPILER)
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<>, decltype(Kokkos::RangePolicy(rp))>);
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(std::move(rp)))>);
-
-  static inline Kokkos::RangePolicy<SomeExecutionSpace> rpses;
-  static_assert(std::is_same_v<Kokkos::RangePolicy<SomeExecutionSpace>,
-                               decltype(Kokkos::RangePolicy(rpses))>);
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<SomeExecutionSpace>,
-                     decltype(Kokkos::RangePolicy(std::move(rpses)))>);
-#endif  // not defined(__INTEL_COMPILER)
-
-  //  RangePolicy
-
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<>, decltype(Kokkos::RangePolicy())>);
-
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(des, i64, i64))>);
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(notEs, i64, i64))>);
-  static_assert(std::is_same_v<Kokkos::RangePolicy<SomeExecutionSpace>,
-                               decltype(Kokkos::RangePolicy(ses, i64, i64))>);
-  static_assert(std::is_same_v<RangePolicyTestExecSpace,
-                               decltype(Kokkos::RangePolicy(es, i64, i64))>);
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(des, i32, i32))>);
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(notEs, i32, i32))>);
-  static_assert(std::is_same_v<Kokkos::RangePolicy<SomeExecutionSpace>,
-                               decltype(Kokkos::RangePolicy(ses, i32, i32))>);
-  static_assert(std::is_same_v<RangePolicyTestExecSpace,
-                               decltype(Kokkos::RangePolicy(es, i32, i32))>);
-
-  // RangePolicy(index_type, index_type)
-
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(i64, i64))>);
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(i32, i32))>);
-
-  // RangePolicy(execution_space, index_type, index_type, Args...)
-
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<>,
-                     decltype(Kokkos::RangePolicy(des, i64, i64, cs))>);
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<>,
-                     decltype(Kokkos::RangePolicy(notEs, i64, i64, cs))>);
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<SomeExecutionSpace>,
-                     decltype(Kokkos::RangePolicy(ses, i64, i64, cs))>);
-  static_assert(
-      std::is_same_v<RangePolicyTestExecSpace,
-                     decltype(Kokkos::RangePolicy(es, i64, i64, cs))>);
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<>,
-                     decltype(Kokkos::RangePolicy(des, i32, i32, cs))>);
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<>,
-                     decltype(Kokkos::RangePolicy(notEs, i32, i32, cs))>);
-  static_assert(
-      std::is_same_v<Kokkos::RangePolicy<SomeExecutionSpace>,
-                     decltype(Kokkos::RangePolicy(ses, i32, i32, cs))>);
-  static_assert(
-      std::is_same_v<RangePolicyTestExecSpace,
-                     decltype(Kokkos::RangePolicy(es, i32, i32, cs))>);
-
-  // RangePolicy(index_type, index_type, Args...)
-
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(i64, i64, cs))>);
-  static_assert(std::is_same_v<Kokkos::RangePolicy<>,
-                               decltype(Kokkos::RangePolicy(i32, i32, cs))>);
-};  // TestRangePolicyCTADs struct
 
 }  // namespace
